整理代码 接收svn接收数据的方法
