+++ CustomSVNBackend/svn/urls_api.py
@@ -2,8 +2,7 @@ from django.urls import path, include
 from rest_framework.routers import DefaultRouter
 
 from svn.views.api_views import FileChangeListLatestExistView, GetFileChangeByRevisionView, GetFileChangesByFilePath
-from svn.views.views import RepositoryViewSet, CommitViewSet, get_latest_revision, receive_svn_data, list_commits, \
-    list_file_changes
+from svn.views.views import RepositoryViewSet, CommitViewSet, get_latest_revision, list_commits, list_file_changes
 
 from svn.views.update_svn import receive_commits
 
@@ -14,8 +13,7 @@ router.register(r'commits', CommitViewSet)
 urlpatterns = [
     path('', include(router.urls)),
     path('repositories/<str:repo_name>/latest_revision/', get_latest_revision, name='get_latest_revision'),
-    # path('receive_svn_data/', receive_svn_data, name='receive_svn_data'),  # 添加receive_svn_data路径
-    path('receive_commits/', receive_commits, name='receive_svn_data'),  # 接收并且储存commits数据.receive_svn_data方法已经被弃用
+    path('receive_commits/', receive_commits, name='receive_svn_data'),  # 接收并且储存commits数据
     path('repositories/<str:repo_name>/commits/', list_commits, name='list_commits'),
     path('repositories/<str:repo_name>/commits/<str:revision>/file_changes/', list_file_changes,
          name='list_file_changes'),
