+++ svn_client/manager.py
@@ -1,18 +1,21 @@
-import os.path
-from urllib.parse import unquote
+import dataclasses
+import sys
+import threading
+import time
 
 import requests
-from config import Config
+from config import SVNClientConfig, Config
+from dc import CommitLogToServerDC
 from endpoints import Endpoints
-from svn_utils import (
-    get_token, get_latest_revision, get_svn_log, parse_svn_log,
-    get_svn_changes, calculate_size, get_latest_svn_revision
-)
 from status_manager import StatusManager
+from svn_utils import get_latest_svn_revision, parse_svn_log2, get_svn_log2, \
+    get_token, calculate_size, get_latest_revision
 
 
 class SVNManager:
-    def __init__(self, status_manager):
+
+    def __init__(self, svn_client_config: SVNClientConfig, status_manager: StatusManager):
+        self.config = svn_client_config
         self.session = requests.Session()
         self.token = get_token(self.session, Config.USERNAME, Config.PASSWORD)
         if not self.token:
@@ -22,107 +25,87 @@ class SVNManager:
             'Content-Type': 'application/json'
         }
         self.status_manager = status_manager
+        self.__DIV_COMMITS_NUM = 5
+
+    def get_start_revision(self) -> int:
+        '''
+        获取起始revision
+        由配置选项传入起始revision，如果起始值为None那么更新则从revision1开始
+        :return:
+        '''
+        start_revision = self.config.START_REVISION
+        existing_latest_revision = self.get_existing_revision()
+
+        if start_revision and 0 < start_revision <= existing_latest_revision and self.config.FORCE_UPDATE:
+            return self.config.START_REVISION
+        else:
+            return existing_latest_revision
+
+    def get_end_revision(self) -> int:
+        '''
+        获取结束revision
+        由配置选项传入结束revision,如果指定了特定的结束revision，则检查是否存在此revision。如果存在则返回，如果不存在则返回当前svn服务器端最大revision
+        :return:
+        '''
+        svn_latest_revision = self.get_existing_revision()
+
+        if self.config.END_REVISION and self.config.END_REVISION < svn_latest_revision:
+            return self.config.END_REVISION
+        else:
+            return get_latest_svn_revision(self.config.REPO_ROOT_URL)
+
+    def get_commits_data(self):
+
+        if self.get_end_revision() - self.get_start_revision() > self.config.COMMITS_SPLIT_NUM:
+            update_to_end_revision = self.get_start_revision() + self.config.COMMITS_SPLIT_NUM - 1
+        else:
+            update_to_end_revision = self.get_end_revision()
+        print(self.get_start_revision(), update_to_end_revision)
+        # 当自定义服务器上的revision版本号大于等于SVN服务器上的revision时，无需上传数据到自定义服务器。也就是此时自定义服务器版本为最新
+        if self.get_start_revision() >= update_to_end_revision:
+            self.config.increment_svn_update_interval()
+            return []
+
+        print(f'获取SVN数据-{self.get_start_revision()}~{update_to_end_revision}')
+        self.config.set_default_svn_update_interval()
+        data = get_svn_log2(self.config.REPO_ROOT_URL, self.get_start_revision(), update_to_end_revision)
+        result = []
+        for d in parse_svn_log2(data):
+            result.append(
+                CommitLogToServerDC(**dataclasses.asdict(d),
+                                    repo_name=self.config.REPO_NAME_CUSTOM_SERVER,
+                                    svn_client_version=self.config.CLIENT_VERSION,
+                                    )
+            )
+        return [dataclasses.asdict(_) for _ in result]
+
+    def update_commits_data(self):
+        '''
+        上传commit函数.也是外部所调用的函数
+        :return:
+        '''
+        self.status_manager.start_upload()
+        try:
+            commits_data = self.get_commits_data()
+            __div_data_size = 0
+            div_data = []
+            for idx, i in enumerate(commits_data):
+                __div_data_size += calculate_size(i)
+                div_data.append(i)
+                if idx == len(commits_data) - 1 or __div_data_size > self.config.MAX_UPDATE_PER_COMMITS_DATA_SIZE:
+                    print(f'上传数据-{div_data[0].get("revision")}~{div_data[-1].get("revision")}')
+                    # 上传数据
+                    response = self.session.post(Endpoints.get_api_url(Endpoints.receive_commits), json=div_data,
+                                                 headers=self.headers)
+                    # 重置分割数据
+                    div_data = []
+                    __div_data_size = 0
+
+        finally:
+            self.status_manager.end_upload()
 
     def get_existing_revision(self):
         latest_revision = get_latest_revision(self.session, Config.REPO_NAME, self.headers)
         return int(latest_revision) if latest_revision is not None else None
 
-    def decode_url(self, url):
-        return unquote(url)
 
-    def update_svn_data(self):
-        self.status_manager.start_upload()
-
-        try:
-            if Config.START_REVISION is not None:
-                start_revision = Config.START_REVISION
-            else:
-                start_revision = self.get_existing_revision()
-                if start_revision is not None:
-                    start_revision += 1
-                else:
-                    start_revision = 1
-
-            print('*' * 50)
-            end_revision = Config.END_REVISION or get_latest_svn_revision(Config.LOCAL_REPO_URL)
-            print(end_revision)
-            if start_revision > end_revision:
-                print(f"No new revisions to update. Current latest revision is {end_revision}.")
-                return
-
-            # Calculate the number of iterations needed
-            iterations = (end_revision - start_revision) // Config.COMMITS_SPLIT_NUM + 1
-            print(f'iterations: {iterations}')
-
-            for i in range(iterations):
-                current_start_revision = start_revision + i * Config.COMMITS_SPLIT_NUM
-                current_end_revision = min(start_revision + (i + 1) * Config.COMMITS_SPLIT_NUM - 1, end_revision)
-
-                log_data = get_svn_log(Config.LOCAL_REPO_URL, start_revision=current_start_revision,
-                                       end_revision=current_end_revision)
-
-
-                commits = parse_svn_log(log_data)
-                if not commits:
-                    print("No new commits to upload.")
-                    continue
-
-                # Get file changes for all commits at once
-                all_file_changes = get_svn_changes(Config.LOCAL_REPO_URL, [commit['revision'] for commit in commits])
-
-                for commit in commits:
-                    # Use the pre-fetched file changes
-                    commit['file_changes'] = all_file_changes[commit['revision']]
-
-                    # Add branch name to commit
-                    if commit['file_changes']:
-                        file_path: str = commit['file_changes'][0]['file_path'].lower()
-
-                        branch_name = None
-                        file_path_replaced = file_path.replace(Config.get_repo_root_url(), '')
-                        if file_path_replaced.startswith('trunk'):
-                            branch_name = 'trunk'
-                        elif file_path_replaced.startswith('branches') or file_path_replaced.startswith('tags'):
-                            split_file_path_replaced = file_path_replaced.split('/')
-                            if len(split_file_path_replaced) > 2:
-                                branch_name = '/'.join(split_file_path_replaced[:2])
-
-                        commit['branch_name'] = branch_name
-                self.upload_commits(commits)
-
-        finally:
-            self.status_manager.end_upload()
-
-    def upload_commits(self, commits):
-        data = {
-            'repository': {
-                'name': Config.REPO_NAME,
-                'url': Config.REPO_URL,
-            },
-
-            'commits': []
-        }
-
-        current_size = calculate_size(data)
-        for commit in commits:
-            commit_size = calculate_size(commit)
-            if current_size + commit_size > Config.MAX_UPLOAD_SIZE:
-                response = self.session.post(Endpoints.get_api_url(Endpoints.svn_receive_svn_data), json=data,
-                                             headers=self.headers)
-                try:
-                    print(response.status_code, response.json())
-                except requests.exceptions.JSONDecodeError:
-                    print(response.status_code, response.text)  # 打印非JSON响应
-                data['commits'] = []
-                current_size = calculate_size(data)
-
-            data['commits'].append(commit)
-            current_size += commit_size
-
-        if data['commits']:
-            response = self.session.post(Endpoints.get_api_url(Endpoints.svn_receive_svn_data), json=data,
-                                         headers=self.headers)
-            try:
-                print(response.status_code, response.json())
-            except requests.exceptions.JSONDecodeError:
-                print(response.status_code, response.text)  # 打印非JSON响应
